Movida la logica de negocio que aun existia en los controller a los service

+++ src/main/java/com/uniovi/services/OffersService.java
@@ -1,10 +1,13 @@
 package com.uniovi.services;
 
+import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
 
 import javax.annotation.PostConstruct;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -19,6 +22,8 @@ import com.uniovi.repositories.UsersRepository;
 @Service
 public class OffersService {
 
+	final static Logger LOG = LoggerFactory.getLogger(OffersService.class);
+
 	@Autowired
 	private OffersRepository offersRepository;
 
@@ -29,7 +34,16 @@ public class OffersService {
 	public void init() {
 	}
 
+	public void addOffer(Offer offer, User activeUser) {
+		offer.setCreator(activeUser);
+		LOG.info("Se ha establecido el creador de la oferta (" + offer.getTitle() + ") a " + activeUser.getEmail());
+		addOffer(offer);
+	}
+
 	public void addOffer(Offer offer) {
+		Date now = new Date();
+		offer.setDate(now);
+		LOG.info("Se ha establecido la fecha de la oferta (" + offer.getTitle() + ") a " + now.toString());
 		offersRepository.save(offer);
 	}
 
@@ -56,8 +70,24 @@ public class OffersService {
 		return offersRepository.findById(offerId).get();
 	}
 
-	public void delete(Offer offer) {
+	public boolean delete(Long offerId, User activeUser) {
+
+		Offer offer = getOfferById(offerId);
+		if (offer == null) {
+			LOG.error("Se ha intentado borrar una oferta que no exite: " + offerId);
+			return false;
+		}
+
+		if (activeUser.getId() != offer.getCreator().getId()) {
+			LOG.error("Se ha intentado borrar una oferta (id = " + offerId + " que no es del usuario: "
+					+ activeUser.getEmail());
+			return false;
+		}
+
 		offersRepository.delete(offer);
+
+		LOG.info("Eliminada offerta con id: " + offerId);
+		return true;
 	}
 
 	public boolean buy(User activeUser, Offer offer) {
@@ -82,7 +112,7 @@ public class OffersService {
 
 		offersRepository.save(offer);
 		usersRepository.save(activeUser);
-		
+
 		return true;
 
 	}
